+++ opo/tophat/defect/RK_OPO_sp.f90
@@ -5,7 +5,7 @@
         USE FFTW3
         IMPLICIT NONE
 
-        real(8) :: x1_r,x2_r,h1_r,hmin_r
+        real(dp) :: x1_r,x2_r,h1_r,hmin_r
 
         ! fft stuff
         ! forward means real space to momentum space
